tech(llm): move databaseBuilders into their own directory for autocompletion
