+++ api/db/database-builder/factory/index.js
@@ -6,6 +6,7 @@ import * as buildDataProtectionOfficer from './build-data-protection-officer.js'
 import * as campaignParticipationOverviewFactory from './campaign-participation-overview-factory.js';
 import * as knowledgeElementSnapshotFactory from './knowledge-elements-snapshot-factory.js';
 import * as learningContent from './learning-content/index.js';
+import * as llm from './llm/index.js';
 import * as poleEmploiSendingFactory from './pole-emploi-sending-factory.js';
 
 const path = dirname(fileURLToPath(import.meta.url));
@@ -49,4 +50,5 @@ export const factory = {
   poleEmploiSendingFactory,
   buildDataProtectionOfficer,
   learningContent,
+  llm,
 };
