+++ api/tests/evaluation/acceptance/application/assessments/assessment-controller_test.js
@@ -888,7 +888,7 @@ describe('Acceptance | Controller | assessment-controller', function () {
             configId: chatDTO.configurationId,
             configContent: chatDTO.configuration,
           };
-          await databaseBuilder.factory.buildChat(databaseChat);
+          await databaseBuilder.factory.llm.buildChat(databaseChat);
           await databaseBuilder.commit();
 
           const promptLlmScope = nock('https://llm-test.pix.fr/api')
